Move graph init functions to overlap module, load read function to seqio module
