+++ kevlar/overlap.py
@@ -170,3 +170,69 @@ def calc_offset(read1, read2, minkmer, debugstream=None):
         print_read_pair(pair, tailpos, debugstream)
 
     return pair
+
+
+def graph_add_edge(graph, pair, minkmer):
+    """
+    Add edge between two nodes in the "shared interesting k-mer" read graph.
+
+    If the edge already exists, make sure that the existing edge matches the
+    edge that would have been added.
+    """
+    tailname, headname = pair.tail.name, pair.head.name
+    if tailname in graph and headname in graph[tailname]:
+        assert graph[tailname][headname]['offset'] == pair.offset
+        if graph[tailname][headname]['tail'] == tailname:
+            assert graph[tailname][headname]['overlap'] == pair.overlap
+        graph[tailname][headname]['ikmers'].add(minkmer)
+    else:
+        graph.add_edge(tailname, headname, offset=pair.offset,
+                       overlap=pair.overlap, ikmers=set([minkmer]),
+                       orient=pair.sameorient, tail=tailname)
+
+
+def graph_init_abund_check_pass(numreads, minkmer, minabund=5, maxabund=500,
+                                logstream=None):
+    """Check whether the k-mer falls within the expected range of abundance."""
+    if maxabund and numreads > maxabund:
+        msg = '            skipping k-mer with abundance {:d}'.format(numreads)
+        print(msg, file=logstream)
+        return False
+    if minabund and numreads < minabund:
+        message = '[kevlar::assemble] WARNING: k-mer {}'.format(minkmer)
+        message += ' (rev. comp. {})'.format(kevlar.revcom(minkmer))
+        message += ' only has abundance {:d}'.format(numreads)
+        out = logstream if logstream is not None else sys.stderr
+        print(message, file=out)
+        return False
+    return True
+
+
+def graph_init(reads, kmers, minabund=5, maxabund=500, logstream=None):
+    """
+    Initialize the "shared interesting k-mer" read graph.
+
+    Iterate through each interesting k-mer, consider every pair of reads
+    containing that k-mer, and determine whether there should be a edge between
+    the pair of reads in the graph.
+    """
+    graph = networkx.Graph()
+    nkmers = len(kmers)
+    for n, minkmer in enumerate(kmers, 1):
+        if n % 100 == 0:  # pragma: no cover
+            msg = 'processed {:d}/{:d} shared novel k-mers'.format(n, nkmers)
+            print('[kevlar::assemble]    ', msg, sep='', file=logstream)
+
+        readnames = kmers[minkmer]
+        if not graph_init_abund_check_pass(len(readnames), minkmer, minabund,
+                                           maxabund, logstream):
+            continue
+
+        readset = [reads[rn] for rn in readnames]
+        for read1, read2 in itertools.combinations(readset, 2):
+            pair = kevlar.overlap.calc_offset(read1, read2, minkmer, logstream)
+            if pair is kevlar.overlap.INCOMPATIBLE_PAIR:
+                # Shared k-mer but bad overlap
+                continue
+            graph_add_edge(graph, pair, minkmer)
+    return graph
