+++ kevlar/tests/test_assemble.py
@@ -12,10 +12,10 @@ import screed
 from networkx import connected_components
 import kevlar
 from kevlar import KmerOfInterest
-from kevlar.assemble import (merge_pair, merge_and_reannotate, load_reads,
-                             graph_init)
+from kevlar.seqio import load_reads_and_kmers
+from kevlar.assemble import (merge_pair, merge_and_reannotate)
 from kevlar.overlap import (OverlappingReadPair, INCOMPATIBLE_PAIR,
-                            calc_offset)
+                            calc_offset, graph_init)
 from kevlar.tests import data_file
 
 
@@ -364,10 +364,10 @@ def test_merge_contained_with_offset_and_error(record7, record12):
     assert 'attempted to assemble incompatible reads' in str(ae)
 
 
-def test_load_reads():
+def test_load_reads_and_kmers():
     """Make sure augmented records are loaded correctly."""
     instream = open(data_file('var1.reads.augfastq'), 'r')
-    reads, kmers = load_reads(instream, logstream=None)
+    reads, kmers = load_reads_and_kmers(instream, logstream=None)
     assert len(reads) == 10
     assert len(kmers) == 7
 
@@ -389,7 +389,7 @@ def test_load_reads():
 def test_graph_init():
     """Test graph initialization."""
     instream = open(data_file('var1.reads.augfastq'), 'r')
-    reads, kmers = load_reads(instream, logstream=None)
+    reads, kmers = load_reads_and_kmers(instream, logstream=None)
     graph = graph_init(reads, kmers, maxabund=None, logstream=None)
 
     # 10 reads in the file, but read16f has no valid connections due to error
@@ -420,7 +420,7 @@ def test_graph_init():
 
 def test_assembly_round2():
     instream = open(data_file('var1.round2.augfastq'), 'r')
-    reads, kmers = load_reads(instream, logstream=None)
+    reads, kmers = load_reads_and_kmers(instream, logstream=None)
     contig, read = reads['contig1'], reads['read22f start=5,mutations=0']
     pair = calc_offset(contig, read, 'AAGTCTCGACTTTAAGGAAGTGGGCCTAC')
     assert pair.tail == read
@@ -432,7 +432,7 @@ def test_assembly_round2():
 
 def test_assembly_contigs():
     instream = open(data_file('AluContigs.augfastq'), 'r')
-    reads, kmers = load_reads(instream, logstream=None)
+    reads, kmers = load_reads_and_kmers(instream, logstream=None)
     contig6, contig7 = reads['contig6'], reads['contig7']
     pair = calc_offset(contig6, contig7, 'AAAGTTTTCTTAAAAACATATATGGCCGGGC')
     assert pair.offset == 50
